+++ server/index.js
@@ -1,5 +1,6 @@
-const messageService = require('../services/messageService')
-const { generateRoomName, SearchListenerOnline, checkIsInRoom } = require('../libs/utility')
+const subscribe = require('./modules/subscribe')
+const publicChat = require('./modules/publicChat')
+const privateChat = require('./modules/privateChat')
 
 module.exports = (server) => {
   const users = new Map()
@@ -26,7 +27,6 @@ module.exports = (server) => {
     socket.on('currentUser', async msg => {
       try {
         socket.data = { ...msg }
-
         const data = {
           userSocketId: socket.id, ...socket.data
         }
@@ -42,6 +42,9 @@ module.exports = (server) => {
 
         // 多網頁連接同帳號判斷
         socket.join(`user${socket.data.id}`)
+        // 進入訂閱的頻道s
+        // const channels = await 
+        // socket.join(channels)
 
         io.emit('users', [...users.values()])
       } catch (error) {
@@ -54,127 +57,9 @@ module.exports = (server) => {
 
     })
 
-    // 未讀通知
-    socket.on('messageNotify', async msg => {
-      try {
-        const unReads = await messageService.searchUnread(io, socket, msg.id)
-
-        socket.emit('messageNotify', unReads)
-
-      } catch (error) {
-        return socket.emit('error', {
-          status: error.name,
-          message: error.message
-        })
-      }
-    })
-
-    // 進入房間
-    socket.on('enterRoom', async msg => {
-      try {
-        const { id, listenerId } = msg
-        let roomName = generateRoomName(id, listenerId)
-        socket.join(roomName)
-        console.log('===========roomName===========', roomName)
-
-        await messageService.createPrivateRoom(id, listenerId, roomName)
-
-        // 先清空 => 後搜尋未讀
-        await messageService.clearUnread(io, socket, msg)
-        const unReads = await messageService.searchUnread(io, socket, id)
-
-        socket.emit('messageNotify', unReads)
-      } catch (error) {
-        console.log(error)
-        return socket.emit('error', {
-          status: error.name,
-          message: error.message
-        })
-      }
-    })
-
-    // 離開房間
-    socket.on('leaveRoom', async msg => {
-      try {
-        const { id, listenerId } = msg
-        let roomName = generateRoomName(id, listenerId)
-
-        socket.leave(roomName)
-
-      } catch (error) {
-        return socket.emit('error', {
-          status: error.name,
-          message: error.message
-        })
-      }
-    })
-
-    // 1on1私聊
-    socket.on('privateMessage', async msg => {
-      try {
-        const { id, listenerId } = msg
-        const roomName = generateRoomName(id, listenerId)
-        // 找到房間內所有連線
-        const clients = io.sockets.adapter.rooms.get(roomName)
-        console.log(clients)
-
-        if (!clients) {
-          console.log('No clients in room')
-          return
-        }
-
-        const { isOnline, listenerSocketIds } = SearchListenerOnline(io, socket, clients, listenerId)
-        console.log(`===========isOnline=============`, isOnline, 'listenerSocketid: ', listenerSocketIds)
-
-        msg.isInRoom = checkIsInRoom(io, socket, clients, listenerSocketIds)
-
-        const message = await messageService.saveMessage(msg)
-        const unReads = await messageService.searchUnread(io, socket, listenerId)
-
-        io.to(roomName).emit('privateMessage', message)
-        io.to(`user${listenerId}`).emit('messageNotify', unReads)
-
-      } catch (error) {
-        console.log(error)
-        return socket.emit('error', {
-          status: error.name,
-          message: error.message
-        })
-      }
-    })
-
-
-
-    socket.on('chatMessage', async msg => {
-      try {
-        const message = await messageService.saveMessage(msg)
-
-        return io.emit('chatMessage', message)
-
-      } catch (error) {
-        return socket.emit('error', {
-          status: error.name,
-          message: error.message
-        })
-      }
-    })
-
-    // 離開公開聊天室事件
-    socket.on('leavePublic', async msg => {
-      const matchingSockets = await io.in(`user${socket.data.id}`).allSockets()
-      const isDisconnected = matchingSockets.size === 0
-
-      if (isDisconnected) {
-        users.delete(socket.data.id)
-
-        socket.broadcast.emit('userDisconnected', {
-          name: socket.data.name,
-          isOnline: 0
-        })
-
-        io.emit('users', [...users.values()])
-      }
-    })
+    subscribe(io, socket)
+    privateChat(io, socket)
+    publicChat(io, socket, users)
 
     socket.on('disconnect', async reason => {
       console.log(reason)
